# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Koch Anomaly Tracker API
  description: |
    Production-grade API for real-time sensor monitoring, anomaly detection, and security management.
    Handles high-concurrency telemetry firehoses and secure access control.
  version: 1.0.0
  contact:
    name: Advanced Agentic Coding Team
servers:
  - url: http://localhost:8080
    description: Local Development Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: P@ssword123
    
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token expiration in seconds
          example: 900

    TokenRefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    PasswordChangeRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    PasswordResetRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string

    PasswordResetCompleteRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8

    AssetSensorReading:
      type: object
      properties:
        readingId:
          type: integer
        assetId:
          type: string
          example: KILN-01
        sensorType:
          type: string
          example: TEMPERATURE
        readingValue:
          type: number
          format: double
          example: 105.5
        uom:
          type: string
          example: C
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [NORMAL, WARNING, CRITICAL]

    PaginatedReadings:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AssetSensorReading'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

paths:
  /api/auth/login:
    post:
      summary: Authenticate user
      tags: [Security]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/refresh:
    post:
      summary: Refresh access token
      tags: [Security]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: New tokens generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token

  /api/auth/me:
    get:
      summary: Get current session identity
      tags: [Security]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Identity found
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized

  /api/auth/password:
    post:
      summary: Change user password
      tags: [Security]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: Password updated
        '401':
          description: Unauthorized

  /api/auth/forgot-password:
    post:
      summary: Initiate password reset
      tags: [Security]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '202':
          description: Reset request accepted (masked response)

  /api/auth/reset-password:
    post:
      summary: Complete password reset
      tags: [Security]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetCompleteRequest'
      responses:
        '200':
          description: Password reset successfully
        '400':
          description: Invalid token or password format

  /api/sensors/readings:
    get:
      summary: Retrieve paginated historical sensor data
      tags: [Telemetry]
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: query
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReadings'
    
    post:
      summary: Ingest new sensor reading (Legacy/Manual)
      tags: [Telemetry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetSensorReading'
      responses:
        '201':
          description: Reading ingested

  /api/sensors/stream/{assetId}:
    get:
      summary: Real-time SSE telemetry stream
      description: |
        Streams real-time sensor readings for a specific asset.
        Supports authentication via Bearer token in 'Authorization' header or 'token' query parameter.
      tags: [Telemetry]
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: query
          description: Alternative for Bearer token for EventSource compatibility
          schema:
            type: string
      responses:
        '200':
          description: Continuous event stream
          content:
            text/event-stream:
              schema:
                type: string
                example: "event: message\ndata: {\"assetId\":\"KILN-01\", ...}"